actu 2

+++ src/ej05.py
@@ -1,14 +1,7 @@
-precio = int(input ("Introduzca el precio sin IVA: "))
+precio = float(input ("Introduzca el precio sin IVA: "))
 
-iva = int(input ("Introduzca el IVA que desea aplicar: "))
+iva = float(input ("Introduzca el IVA que desea aplicar: "))
 
 total = precio * iva
 
 print ("El precio del articulo con IVA es", total)
-
-
-
-
-
-
-
